Keep tests close to components
