+++ src/components/Caption/__tests__/Caption.spec.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import Caption from '../Caption';
+
+describe('<Caption />', () => {
+    it('should render a filter caption', () => {
+        const wrapper = shallow(
+            <Caption>
+                This is a filter caption
+            </Caption>
+        );
+
+        expect(wrapper).toMatchSnapshot();
+    });
+    it('should not render an empty caption', () => {
+        const wrapper = shallow(
+            <Caption />
+        );
+
+        expect(wrapper).toMatchSnapshot();
+    });
+});
