+++ src/components/FilterInfo/__tests__/FilterInfo.spec.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import FilterInfo from '../FilterInfo';
+
+describe('<FilterInfo />', () => {
+    it('should render a default filter info tooltip', () => {
+        const wrapper = shallow(
+            <FilterInfo
+                information="Some filter information"
+            />
+        );
+
+        expect(wrapper).toMatchSnapshot();
+    });
+
+    it('should not render anything if prop is empty', () => {
+        const wrapper = shallow(
+            <FilterInfo />
+        );
+
+        expect(wrapper).toMatchSnapshot();
+    });
+});
