+++ src/components/LoadingIndicator/__tests__/LoadingIndicator.spec.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import LoadingIndicator from '../LoadingIndicator';
+
+describe('<LoadingIndicator />', () => {
+    it('should render a default loading indicator', () => {
+        const wrapper = shallow(
+            <LoadingIndicator />
+        );
+
+        expect(wrapper).toMatchSnapshot();
+    });
+
+    it('should render a custom loading indicator', () => {
+        const wrapper = shallow(
+            <LoadingIndicator
+                size="12"
+            />
+        );
+
+        expect(wrapper).toMatchSnapshot();
+    });
+});
